Added Buildings, Floors, Rooms, and GeospatialSpace (join model)

+++ db/migrate/20240906132628_remove_unused_columns_from_better_together_pages.rb
@@ -3,7 +3,7 @@
 # Cleans up unused columns on pages table
 class RemoveUnusedColumnsFromBetterTogetherPages < ActiveRecord::Migration[7.1]
   def change
-    remove_column :better_together_pages, :language
-    remove_column :better_together_pages, :published
+    remove_column :better_together_pages, :language  if column_exists?(:better_together_pages, :language)
+    remove_column :better_together_pages, :published if column_exists?(:better_together_pages, :published)
   end
 end
